Adjust for gcc error message changes.
[Not reviewed.]

From gcc 4.3.4 to 4.7.2, we get some changes in the error messages
produced when the target C compiler processes the C code we produce
for this test.  Adjust the .prediff and .bad so that the latter can
match for both gcc compiler versions (and probably the ones between
those also, although I didn't check that).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23387 3a8e244f-b0f2-452b-bcba-4c88e055c3ca