Promote r23383 to release branch.
Original message:

Add a warning if GASNet gemini or aries conduit are used and hugepages is loaded

[reviewed by Greg]

Add a script that checks if a craype-hugepages module is loaded while using
GASNet gemini or aries conduits and prints a warning that it is not supported.

Also print a warning if comm-ugni is used and craype-hugepages is not loaded.

The new script is called in Makefile.exe right before calling the C compiler
and doesn't stop the compilation if there is a warning.

The script is also called as the last thing that happens in the GASNet build.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/release/1.9@23388 3a8e244f-b0f2-452b-bcba-4c88e055c3ca