Hoist Stmt out as an abstract base class of the remaining Stmt<?> classes.
Created the abstract base class Stmt, and made other Stmt types inherit from
it.  This allows the definition of the predicate IS_STMT to be simplified using
a virtual predicate function isStmt().

This could also be done using RTTI, but the LLVM builds specifically disable
RTTI as an optimization, so that is ruled out for now.  The virtual function
approach is probably cleaner anyway.


DETAILS

    compiler/include/stmt.h
Created the base class Stmt and populated it with common virtual functions.
The constructor just initializes the Expr base-class sub-object.  The class is
abstract due to the presence of the pure virtual implementations for
replaceChild() and verify().
Added the predicate isStmt() which is true for the Stmt class and (by default)
all of its descendents.
Now the various Stmt types inherit from the Stmt base class.
Added the "virtual" keyword to replaceChild() and verify() in derived classes,
to make them more self-documenting.

    compiler/include/expr.h
Removed the IS_EXPR() macro as dead code, and replaced the IS_STMT() macro with
an isStmt() predicate in the Expr class interface.  In the revision, isStmt()
is true for all descendents of Stmt.  The previous macro did not include
ExternBlockStmt, but this was most likely an oversight.
Added the "virtual" keyword to declarations of replaceChild() and verify() in
derived classes, to make them more self-documenting.

    compiler/AST/view.cpp
    compiler/AST/expr.cpp
Replaced uses of the IS_STMT() macro with the isStmt() predicate.

    compiler/AST/stmt.cpp
Added the trivial destructor Stmt::~Stmt().
Constructors in derived classes now perform base-class initialization through
their parent class (Stmt).

[tested on / (std)]
[reviewed by noakes]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23390 3a8e244f-b0f2-452b-bcba-4c88e055c3ca