Define a simple abstract Visitor base class for AST nodes
I have defined a simple base class to support pre-order traversals of
AST nodes.  Tom thinks this will cover the dominant use cases but we
might want to add support for other traversals at a later time.

This interface a simple visit() for the sub-types that never have
children and an enter/exit pair for those that do.  The enter()
version returns a bool that indicates whether to contine the
traversal or not.

[Compiled for std. No interaction with LLVM. No uses and so did not run paratest]
[reviwed by Tom]





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23391 3a8e244f-b0f2-452b-bcba-4c88e055c3ca