[Patch] Add the "accept" side of the AstVisitor
Insert the accept() side of the AstVisitor pattern in to the
concrete sub-classes of BaseAST.

The signature is the same for every subclass.  There are two styles of
implementation

1) If the sub-class has no children fields then just call

   visitor->visit(this);

2) If the sub-class has children then pattern is

   if (visitor->visitEnter(this) == true) {

      <traverse children>

      visitor->visitExit(this);
   }

The <traverse children> step mirrors the traversal done in the
macro AST_CHILDREN_CALL()

For this checkin there are no uses of this API but this API has
been tested in three Visitor applications that will be checked in
shortly.


[Compiled for std.  No interaction for LLVM. No uses and so did not run paratest]
[Revieved Tom]



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23397 3a8e244f-b0f2-452b-bcba-4c88e055c3ca