Make an empty struct nonempty.
In r23405 we added a component of the task private data to hold comm
layer information.  This is chpl_comm_taskPrvData_t, a struct typt.  For
comm=none we don't need any comm layer info, so we made the struct type
empty.  Unfortunately empty struct types aren't legal C (c99 standard,
6.2.5 (20)).  The gcc compiler lets this pass, but the PGI and Cray
compilers bark at it.  To fix this, add a dummy element to the struct
for the comm=none case.  (I tried making it a 0-length array, but the
PGI compiler barked at that, too.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23411 3a8e244f-b0f2-452b-bcba-4c88e055c3ca