Split view.cpp into three separate files as a first step to breaking out the various loggers
I have been doing some work to develop a variation of dump_ast() that
produces output that exposes more of the inner detail of the Node
structure of the AST at the cost of being less "reader friendly".
This version is aimed at supporting the early development of the
interpreter but may be of help to other developers.

This new version will sit alongside the current versions and be invoked
with new command line flags.

Along the way I 
   1) Developed a basic implementation for a  traditional AST "Visitor"
      that is adequate for these AST viewers. The Visitor may evolve as
      newer use cases are identified.

   2) Recoded the dump_ast() routine using this pattern.  It's certainly
      a little more verbose but, I think, a little better organized and
      slightly easier to follow.

   3) Cloned/modified (2) to develop the "node" format (in progress)

   4) Recoded the html_view() routine using this pattern

The revised implementations of dump_ast() and html_view() are intended
to produce exactly the same output as the current versions.  I have
verified that they do so for a sequential version of Triad where

   dump_ast()  produces 1.7M lines of text
   html_view() producrs 4.5M lines of text



The challenge now is to submit the code changes in an effective manner.
Tom has agreed to review the changes involved but we're going to do it
in a moderate number of incremental updates.


This first submit splits view.cpp in to three files
   1) AstDump.cpp            The code for the human readable form
   2) AstDumpToHtml.cpp      The code for the HMTL form
   3) view.cpp               The remaining code

There is some duplication of code after this split that will be
addressed as the refactoring proceeds.

I have also migrated the few lines of dump_html initialization code from
runpasses.cpp to AstDumpToHmtl.cpp.


Compiled on STD.
Tested   on mac
Reviewed by Tom.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23412 3a8e244f-b0f2-452b-bcba-4c88e055c3ca