Step 2 of cleaning up loggers :- convert to class objects
This patch packages the functions and file static variables in
AST/AstDump.cpp and AST/AstDumpToHtml.cpp as simple C++ classes.

This pass preserves the existing interface between runpasses.cpp
and view.cpp i.e. one based on function calls.

The HTML dumper had an implied singleton to handle updates to the
master index.html file and then instances for each module.  The
AST dumper had implied instances for each module.

This interface is exposed as a small set of static class methods
which might be thought of as "factory" methods.

The remaining changes are primarily name changes to convert the
static functions to private methods and then conversion of
certain arguments to references to instance state.

Migrated the interface to these classes from view.h to the new
class headers.


Compiled: STD and LLVM
Tested:   Compared output for Sequential Triad
Reviewed: Tom




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23416 3a8e244f-b0f2-452b-bcba-4c88e055c3ca