Adapt to varying aprun depths.
[Not reviewed, seems to simple to need it.]

I just did some runs on a Cray system with other than 16 "cores"
(actually hyperthreads) using the aprun launcher, and got a mismatch
on the aprun command line.  Here I've made changes so that only the
presence of the aprun depth (-d) parameter matters, not the value
given to it.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23417 3a8e244f-b0f2-452b-bcba-4c88e055c3ca