Resolve the return type of extern functions.
Make the compiler pay attention to declared return types, particularly when
these are attached to extern functions.  This involves adding a return value
variable declaration to the (stub) body of an extern function and resolving
the return type normally.


DETAILS

    test/statements/lydia/externMethodCallPerf.chpl
This change discovered a syntax error in the declaration:
  extern proc blah2(int): int;
In the formal argument list, Chapel requires a variable name and colon before a
type name.

    test/classes/bradc/overloadMethods/v2NoReturnValueButType.good
This test now succeeds, because the return-value variable is properly initialized.
Note that there is a special case in isDefinedAllPaths, that treats a call to
"halt" as if it defined the return value variable (i.e. in addition to the
initialization).

    test/compflags/bradc/minimalModules/minModFnRefArg.chpl
    test/compflags/bradc/minimalModules/minModDeclPrint.chpl
Since extern prototypes are now resolved, the tuple record must be defined if
an extern signature uses varargs.

    test/users/ferguson/bulkerr.good
The change correctly logs a warning for this test code.
  proc myread(inout args ...?k):bool {
    for param i in 1..k {
      writeln(args(i));
    }
  }
More cases are flagged as having a specified return type, so this case now gets
caught.

    test/functions/hilde/retTypeNoRet.chpl
    test/functions/hilde/retTypeNoRet.good
Added a new test to show that a function having a declared nonvoid return type
and no return statement produces a warning.

    compiler/include/symbol.h
Added a comment on how the retType field is used.

    compiler/AST/type.cpp
Added a flag so gUnknown is treated as a type variable.

    compiler/AST/astutil.cpp
Moved the definition of prune2 down to match the declaration order (bottom-up)
in the rest of the file.
Added the routine removeVoidMoves to remove assignments of _void from the tree.

    compiler/passes/normalize.cpp
Treat variables of type 'void' as always defined, to avoid use-before-def
errors on them.
Moved the insertion of a void return down past the test that a function is
already normalized.

    compiler/passes/insertLineNumbers.cpp
Never insert line numbers on functions declared to be "extern".

    compiler/passes/buildDefaultFunctions.cpp
Disabled the assert that the initCopy function being built is not a tuple.  We
need the function built for the minimal modules cases that use varargs:
 test/compflags/bradc/minimalModules/minModFnRefArg.chpl
 test/compflags/bradc/minimalModules/minModDeclPrint.chpl

    compiler/resolution/wrappers.cpp
Simplified the case where a wrapper has a void return type.  With this change,
the same code works with extern functions as with normal ones.

    compiler/resolution/functionResolution.cpp
Found and removed a wrapper move that slipped in with Chris Wailes' partial
instantiation work.  His patch was based on a version that preceded the final
assign-to-ref patch.
Added an assert that no more wrapper moves exist in the tree.
TODO: Following testing, this test should be removed.
Allow the AST to contain ('move' lhs rhs) where lhs and rhs are both of void
type and lhs is the return value variable.  This permits the initialization
  var ret:void = _void;
as a specific case of the new way of handling declared return value types.  The
new way is to insert the declaration
  var ret:<declared_type> = ('init' <declared_type>);
at the top of the function.
Split out the resoveReturnType() function, so it can be called simply (after
resolveBody) on extern functions.
Now resolve the body and return type of extern functions.

    compiler/optimizations/narrowWideReferences.cpp
Extern routines are always narrow, so don't add them to the list of
functions to be narrowed.


[tested on / (std)]
[tested on / (no-local)]
[reviewed by noakes]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23420 3a8e244f-b0f2-452b-bcba-4c88e055c3ca