Rework error detection on forms of new-expression.
Existing detection of invalid forms of 'new' expressions is performed in
normalize.cpp.  It assumes that all 'new' expression have been converted into
constructor calls at that point.  This won't work if we want to insert new
expressions later or we want to call a constructor using a type expression that
is more complex than just a type name.

This patch removes the old, lame test and replaces it with a check (in
checkNormalized) that ensures that 'new' expressions look like constructor
calls.  The predicate isTypeExpr() was moved from functionResolution.cpp to
astutil.cpp and made public so it can be used in both functionResolution.cpp
and checkNormalized.cpp.

The error message for new-expressions that do not appear to be constructor
calls was updated to be more helpful (hopefully), and the .good files for the
affected tests were updated.

Since we cannot know before resolution whether a function call names a
constructor, that case is deferred to resolution.  A new routine, resolveNew()
was added to handle that case.


DETAILS

Index: test/distributions/sungeun/misc/no_use_BlockDist.good
We defer until resolution whether the symbol Block represents a type in this
context, so the error changes from "invalid use of 'new'" to an unresolved
symbol error.
In anticipation of being able to call any constructor as the argument of a new
expression, this error message is consistent.  
The intentional error in the test case is a missing "use".  Therefore the Block
type is undefined and an attempt to call its constructor with a specific set of
arguments fails.

Index: test/classes/diten/newArray.good
The same structural error is detected, the but error message was updated to
give a more detailed description of the problem.

Index: test/functions/bradc/resolveGeneric.good
The expression x.type(); shouild be interpreted as a constructor call.  It was
being caught in normalization with the error "invalid use of 'new'".
checkPrimNew() now lets this pass, but it now fails with "illegal call of type" in
resolution.

Index: compiler/include/astutil.h
The isTypeExpr() predicate is now a public symbol defined in astutil.cpp.

Index: compiler/AST/astutil.cpp
The isTypeExpr() predicate was moved here from functionResolution.cpp.

Index: compiler/passes/normalize.cpp
The original test for invalid use of new was removed.  It tested that the
argument of a 'new' expression is a call and that the unresolved symbol in it
does not begin with an alphabetic character.
Presumably, that was to ensure that the function called looked like a
constructor call.
Added a header comment to processSyntacticDistributions().

Index: compiler/passes/checkNormalized.cpp
Pushed the existing functionality of checkNormalized() down into
checkFunctionSignatures(), and added a new subroutine checkPrimNew() to perform
structural checks on 'new' expressions.

Index: compiler/resolution/functionResolution.cpp
Moved the isTypeExpr() predicate to astutils.cpp.
Added resolveNew() to handle new primitives that are not converted in the
normalize pass.  This will allow expressions of the form
  <type-expression> ( <args> )
and 
  <constructor-call>
to be resolved as permissible arguments of a 'new' expression.  In this patch,
only the second form is supported, and the first will result in an error.

[tested on / (std)]
[tested on / (no-local)]
[reviewed by lydia]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23421 3a8e244f-b0f2-452b-bcba-4c88e055c3ca