Shuffle AST logging code to emphasize the intent to dispatch on AST classes
The bulk of these changes are modest code motions and/or code duplications
that emphasize the notion that the dumpers are dispatching on the types
of BaseAST nodes.  This is a prelude to rewriting as visit() methods.



Compiled on STD and LLVM
Tested:    running diffs on Sequential Triad
Reviewed:  Tom



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23424 3a8e244f-b0f2-452b-bcba-4c88e055c3ca