Further code motions to set up for the migration to AstVisitor
The bulk of these changes are modest code motions and/or code duplications
that emphasize the notion that the dumpers are dispatching on the types
of BaseAST nodes.  This is a prelude to rewriting as visit() methods.

In AstDump.cpp
   1) Make the header() function return	a bool which plays the same role as
      the bool that will be returned by	visitEnter()

   2) Migrate the special tests	     in write() that	return early into header()

   3) Flatten an else-if to further emphasize the class dispatch

In AstDumpToHtml.cpp
   1) Introduce	the header/footer pair

   2) Flatten an else-if to further emphasize the class dispatch



Compiled on STD
Tested:    running diffs on Sequential Triad
Reviewed:  Tom



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23426 3a8e244f-b0f2-452b-bcba-4c88e055c3ca