Adding two futures that exercise different expressions of an unclear error message for function creation [reviewed by Greg]
The declarations

proc foo();

and

proc foo(): <return-type>;

both result in internal errors within the compiler, with the former occurring
in checkAST and the latter in functionResolution.  Since the messsages
themselves do not indicate that the problem is with failing to define the
function's body, I consider this a bug, albeit a low-priority one.  When the
problem is resolved, it ought to be possible to merge these tests into a single
one which exercises both cases.  I separated them because the presence of one
hides the error message of the other.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23432 3a8e244f-b0f2-452b-bcba-4c88e055c3ca