Final conversion for AST dumpers to use Visitor pattern
The final conversion to the visitor pattern.

1) Replace header() and footer() with the appropriate set of
visitEnter(), visitExit() and visit() functions.


2) Replace the top level call to write(BaseAst*) with a call
to the accept function for the node


3) Remove the "hack" function that was used to jump through
to AST_CHILDREN_CALL() macro


4) Introduce a trivial base class AstLogger to serve at the
parent to the new form of AstDump and AstDumpToHtml.  A number
of the AstDump and AstDumpToHmtl visitor methods had the
default implementation.  The parent helps to reduce visual
clutter in the child classes.


Compiled: STD and LLVM
Tested:   Verify output for Sequential Triad
Review:   Tom




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23434 3a8e244f-b0f2-452b-bcba-4c88e055c3ca