Patch to fix our aries/gemini memory unmapping issue when pshm is disabled
[contributed by the GASNet team]

This patch is related to r23237 which fixed memory registration problems for
the gemini conduit. That patch changed a use of mmap to gasneti_huge_mmap when
huge pages are available and pshm is disabled. Similarly this patch changes
munmap to gasneti_huge_munmap.

This patch fixes the issue with the aries conduit (and likely gemini as well
though that wasn't confirmed) which caused programs to fail with this error:
*** FATAL ERROR: munmap(0x00002aaa eae00000,13420724224) failed: Invalid
argument

This patch and r23237 are contained in the gasnet source code now so we will
not need to reapply these patches when we upgrade versions next.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23436 3a8e244f-b0f2-452b-bcba-4c88e055c3ca