Simplify business logic for "logging" SymExpr and delete deprecated get_constant()
Supporting discussion:

I was trying to understand the significance of the code to display a
SymExpr as an additional hint to understanding the data model for SymExpr.  



1) The production code for logging a SymExpr to plain text is

void AstDump::visit(SymExpr* node) {
  if (isBlockStmt(node->parentExpr)) {
    newline();
  }

  if (VarSymbol* c = get_constant(node)) {
    if (c->immediate) {
      const size_t bufSize = 128;
      char         imm[bufSize];
      char         buff[bufSize];

      snprint_imm(imm, bufSize, *c->immediate);
      sprintf(buff, "%s%s", imm, is_imag_type(c->type) ? "i" : "");

      write(buff);
    } else {
      write(c->name);
    }

  } else {
    writeSymbol(node->var, false);
  }
}

and the code for AstDumpToHtml is structurally similar.  


A casual glance at the code suggests it does "something" to a SymExpr* 
that might return a VarSymbol*.  If it obtains a VarSymbol it determines
if that value is an immediate or not and takes the appropriate action
otherwise it calls writeSymbol on the node's <var> slot.

I find this code to be a little more opaque than it could be and so I
checked the definition of get_constant().  It turns out that these are
the only two calls to 

VarSymbol* get_constant(Expr* e) {
  if (e) {
    if (SymExpr* l = toSymExpr(e)) {
      if (VarSymbol* v = toVarSymbol(l->var)) {
        if (v->immediate != 0)
          return v;
      }
    }
  }

  return 0;
}


We can see that get_constant is only ever called on a SymExpr*
and that it will only return a non-null VarSymbol* if the var slot 
references an immediate value.




2) This means we can simplify these two functions to become

void AstDump::visit(SymExpr* node) {
  if (isBlockStmt(node->parentExpr)) {
    newline();
  }

  if (VarSymbol* c = get_constant(node)) {
    const size_t bufSize = 128;
    char         imm[bufSize];
    char         buff[bufSize];

    snprint_imm(imm, bufSize, *c->immediate);
    sprintf(buff, "%s%s", imm, is_imag_type(c->type) ? "i" : "");

    write(buff);

  } else {
    writeSymbol(node->var, false);
  }
}

VarSymbol* get_constant(SymExpr* e) {
  VarSymbol* retval = 0;

  if (VarSymbol* v = toVarSymbol(e->var)) {
    if (v->immediate != 0)
      retval = v;
  }

  return retval;
}






3) I believe this business logic is more straight forward if
we simply fold these two functions together i.e.

void AstDump::visit(SymExpr* node) {
  Symbol*    sym = node->var;
  VarSymbol* var = toVarSymbol(sym);

  if (isBlockStmt(node->parentExpr)) {
    newline();
  }

  if (var != 0 && var->immediate != 0) {
    const size_t bufSize = 128;
    char         imm[bufSize];
    char         buff[bufSize];

    snprint_imm(imm, bufSize, *var->immediate);
    sprintf(buff, "%s%s", imm, is_imag_type(var->type) ? "i" : "");

    write(buff);

  } else {
    writeSymbol(sym, false);
  }
}

and then eliminate get_constant().




This minor change makes it clearer that the two logging functions

a) inspect node->parentExpr to determine whether to generate a
line break

b) then operates on the node->var property

It is also interesting to note that this code implicitly assumes that
node->var is non-NULL; it could not be an "immediate" it if were not
and it can be determined that writeSymbol will seg-fault on non-NULL.



Longer term there is a small opportunity to extract the common control
logic between AstDump and AstDumpToHtml for this case but I am not
pursuing this at this time.





It would not take much of a nudge for me to re-write this as some
variant on

void AstDump::visit(SymExpr* node) {
  Symbol* sym = node->var;

  if (sym != 0) {
    VarSymbol* var = toVarSymbol(sym);

    if (isBlockStmt(node->parentExpr)) {
      newline();
    }

    if (var != 0 && var->immediate != 0) {
      const size_t bufSize = 128;
      char         imm[bufSize];
      char         buff[bufSize];

      snprint_imm(imm, bufSize, *var->immediate);
      sprintf(buff, "%s%s", imm, is_imag_type(var->type) ? "i" : "");

      write(buff);

    } else {
      writeSymbol(sym, false);
    }

  } else {
    USR_FATAL("SymExpr->var is unexpectedly NULL");

  }
}

but (a) I am unclear how popular these kinds of asserts() are
and (b) I am not sure what the current practice is for throwing
this kind of exception.


Compiled: STD
Tested:   Validate output on Sequential Triad
Reviewed: Vass



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23437 3a8e244f-b0f2-452b-bcba-4c88e055c3ca