Making pidigits-pretty2 into a future
Our memory leaks for the full-suite testing took a sharp uptick
recently due to this test which was committed with a comment
saying that it had a memory leak, yet I hadn't really done the
analysis to think about how significant it would be.  This commit
makes the test into a future, pending some better support for
handling such cases.  The core challenge is how to define a binary
operator on a type that requires cleanup to the temp used to store
the result:

  proc *(a: T, b: T) {
    var res: T;
    init(res);   // who will clean this up?
    res = ...compute on a and b
    return res;  // whoops, we've just leaked res
  }




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23439 3a8e244f-b0f2-452b-bcba-4c88e055c3ca