Make it harder to abuse function_exists(), and fix up dangerous calls.
The existing implementation of function_exists() is flexible, but unsafe.  By
default, it did not require the selected function to match by the number of
arguments, and if the desired type of any given argument was left off, it would
return a match.

The use of function_exists() has migrated to being a test for the existence of
certain special functions with well-known argument types.  As such, it is
better that function_exists() test for exact matches and fail otherwise.

There was one place where function_exists() was expected to match an argument
of a more general type.  Support for an arbitrary match with dtAny was added to
handle this case.

TODO: Document type_match() and the effect of the dtAny match on the interface
to function_exists().
TODO: Can the right side of the <~> operator be stated precisely a priori?

[tested on / (std)]
[reviwed by sungeun]



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23443 3a8e244f-b0f2-452b-bcba-4c88e055c3ca