Simple cleanup runpassses.cpp
1) Integrate main/passlist.h directly in to main/runpasses.cpp


2) Pulled some global state variables from driver.{h,cpp} in to
main/runpasses.{h,cpp}


3) Determined that some files failed to compile because they relied
on driver.h but not runpasses.h

   a) Moved main/runpasses.h in to include/runpasses.h
   b) Updated required files


4) Removed the macros FIRST and LAST.

FIRST isn't being used in a meaningful way right now and requires
that the two passList loops code jump around the first (zero-th?) slot.
I think FIRST is a legacy issue that was done to arrange for the first
pass to be index as "1".



LAST was used to tag the last entry in

static PassInfo sPassList[] = { };

so that the loops over sPassList[] knew when to stop.

Changed to the more robust cliche of

   size_t passListSize = sizeof(sPassList) / sizeof(sPassList[0]);

   for (size_t i = 0; i < passListSize; i++) ...


A longer term direction might be to make "objects" out of each pass 
description and have the passList be a dynamically-sized vector of
these objects.  Ignored this opportunity for this quick pass.



5) Rearranged the order of functions to support reader's flow when reading
the code and did some trivial syntax tweaks.




Compiled: STD and LLVM with clang on mac and STD on chap01 with gcc.
          gcc issued an extra signed vs unsigned error that clang ignored. Fixed.

Tested:   verified that both log dumpers have same behavior on Mac
          ran paratest on chap*

Reviewed: Tom




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23449 3a8e244f-b0f2-452b-bcba-4c88e055c3ca