Do verbose barrier output only with CHPL_COMM_DEBUG and --verbose.
[Reviewed by Sung.]

This makes the verbose barrier output for comm=gasnet only happen if you
build with CHPL_COMM_DEBUG and run with --verbose.  Whether or not this
output appeared used to be controlled by {start,stop}VerboseComm(), but
we decided that was inappropriate because conceptually, verbose comm is
about what the user program is doing, but how and when the comm barrier
calls are made isn't under the user's control.  So we've changed what
controls this output, to make it more explicit that it has to do with
comm layer debugging.

The specific file changes are as follows.

runtime/make/Makefile.runtime.include:
  If CHPL_COMM_DEBUG is set in the environment, and is thus a make(1)
  variable, also define it to the preprocessor.

runtime/src/comm/gasnet/comm-gasnet.c:
  Produce the comm barrier entry message iff CHPL_COMM_DEBUG is set and
  the program is run with --verbose.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23452 3a8e244f-b0f2-452b-bcba-4c88e055c3ca