Remove superfluous CHPL_COMM_DEBUG check.
[Reviewed by Sung.]

This removes a superfluous CHPL_COMM_DEBUG check in make/Makefile.base.
The only result of this check is to set CHPL_MAKE_COMM_EXTRAS to one
thing or another, and the string CHPL_MAKE_COMM_EXTRAS occurs nowhere
else in the repository, so this can't be needed code.  (I suspect that
the need for this code disappeared when we added the CHPL_COMM_DEBUG
check to third-party/gasnet/Makefile.include, but we failed to remove it
at the time.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23453 3a8e244f-b0f2-452b-bcba-4c88e055c3ca