Collect additional time info for runpasses (phase 1 of 2)
I am updating the way runpasses() collects timing information for each
pass.  The existing functionality is unaffected but additional detail
will be extracted and some additional summary information will be available
at the end of the run.

Doing this in two passes to simplify the code review.


1) Convert runPass() into a function that takes the current passNumber
rather than the components of the Pass data structure.


2) runpass() uses some local "statics" and some slightly convoluted
logic to track state that is logically maintained across distinct
iterations.  Move this state to the "top level" i.e. the function
statics are moved to file scope, and some of the business logic
is moved from runpass() to a spot below the driver loop in runPasses().


Compiled: STD clang on Mac and STD gcc on chap01
Tested:   Verified that print-passes generates the same output
Reviewed: Elliot




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23458 3a8e244f-b0f2-452b-bcba-4c88e055c3ca