Refactor .gitignore into readable, intuitive format.
Previously, .gitignore was a raw dump of `git-svn show-ignore`. We are now
moving towards using git (coming soon!), so this makes the .gitignore more
manageable and concise. For now, there is a single top level .gitignore. It
might make sense to create a separate test/.gitignore in the future for all the
obscure test ignore rules.

r: kyle-b

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23468 3a8e244f-b0f2-452b-bcba-4c88e055c3ca