Replace the global map resolvedFns with a flag and implement the isResolved() method on function symbols.
The code to avoid resolving a function multiple times involved adding it to a
global map: FnSymbol* --> bool.  Using a flag seemed more efficient, since
whether or not a function is resolved is a property of the function itself.

The time for the resolve pass on test/studies/jacobi.chpl dropped by .014s,
averaged over 5 runs.  This difference is greater than 2 standard deviations,
so it is statistically significant.

Added checkRetTypeMatchesRetVarType(), to show that the contents of the retType field is
just a cached version of the return value variable's type after resolution.  Removed
checkNoWrapperMoves() as it has served its purpose:  All code that used to add add wrapper
moves has been removed from the compiler. 


DETAILS

    compiler/include/symbol.h
    compiler/AST/symbol.cpp
Added the predicate FnSymbol::isResolved().

    compiler/include/flags_list.h
Added the "resolved" flag.
Reordered a few entries to put them back in alphabetical order.

    compiler/include/baseAST.h
For all symbol types, the flags were being copied in both the
DECLARE_SYMBOL_COPY() and in copyInner().  It's better to have the copy in
copyInner, since this allows copyInner to override the verbatim copy of all
flags where this is needed.

    compiler/resolution/functionResolution.cpp
Replaced the resolvedFns map with setting and querying the new FLAG_RESOLVED.
In buildVirtualMaps(), split out the tests for which functions are added to the
virtual maps, and added comments to each test.

    compiler/main/checks.cpp
Added checkRetTypeMatchesRetVarType().
Removed checkNoWrapperMoves().


TODO: Replace DECLARE_SYMBOL_COPY() with a literal implementation of copy()
that in each class just calls the copy constructor for that class and returns
the result.

[tested on / (std)]
[tested on release/examples/primers/ (verify)]
[reviewed by noakes]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23470 3a8e244f-b0f2-452b-bcba-4c88e055c3ca