Remove CHPL_RT_NUM_THREADS_PER_LOCALE=2 from gasnet+qthreads testing
[ok'd by Greg]

We have been setting CHPL_RT_NUM_THREADS_PER_LOCALE=2 for qthreads+gasnet
testing for a long time. Recently when full gasnet+qthreads testing was turned
on it was discovered that this was the source of several timeouts. Tracing the
logs and various previous names of this file this was originally set in r20083
to limit the number of workers that qthreads would use per locale since there
are multiple locales running on the same machine with gasnet local spawning. We
have since enabled qthreads oversubscription, so we should be able to stop
setting this. 

Tested with on multilocale/ 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23477 3a8e244f-b0f2-452b-bcba-4c88e055c3ca