Fix up CHPL_TARGET_ARCH for some configurations
[reviewed by sung]

This addresses some of the problems that came up when testing under
different configurations (pgi/intel/cray/...). The primary issue was tests
that threw --fast in their .compopts files were getting a warning that
CHPL_TARGET_ARCH was not set. This is a good warning that we want for end
users, but is needless noise in the test system.

The solution I went with was to add a new value for CHPL_TARGET_ARCH:
'none'. When this is set, --specialize becomes a no-op and the compiler
will not throw any warnings about it. The testing system now sets
CHPL_TARGET_ARCH=none by default and unsets the value for performance and
--fast testing.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23479 3a8e244f-b0f2-452b-bcba-4c88e055c3ca