Rename AstVisitor methods to avoid overiding of overloaded virtual functions
Changed the names of every visit* function in AstVistor to provide a distinct
name for each type.

Corresponding changes in the accept() functions expr.cpp, stmt.cpp, symbol.cpp,
type.cpp to call the appropriate method.


Compiled: STD on clang and STD and LLVM on chap01 (gcc)
Tested:   Compared output of AST dump functions for sequential triad
Reviewed: Tom



More discussion

There is potential for concern when one overrides an overloaded virtual
function e.g. if one defines

  virtual bool visitEnter(AggregateType* node);
  virtual bool visitEnter(EnumType*      node);

in a base class and then does

  virtual bool visitEnter(EnumType*      node);

in a derived class.  Under some circumstances and with some compilers it
can be "hard to tell" which virtual method will be selected when these
methods are called using pointers with different derived types.



GCC is reported to be less sophisticated about this in general.  For reasons
that I can't explain GCC builds work fine on a number of configurations but
fail on the Chap machines for LLVM even using gcc 4.7.*




Several sites recommend never relying on overloading a virtual function
although one site specifically mentioned the common style for the Visitor
pattern as a common exception to this guideline.

One proposed "work around" that might have provided safety and calmed the
compiler did not work.  Hence I changed the names to eliminate the overloads.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23482 3a8e244f-b0f2-452b-bcba-4c88e055c3ca