Remove removeNilTypeArgs(), for readability.
The step removeNilTypeArgs() in pruneResolvedTree() used to express arguments
of type dtNil by removing them from the tree.  This was apparently done to
avoid having to deal with nil type objects explicitly when ensuring that formal
and actual types match.

However, the actual test (that an argument of nil type can be promoted to any
class type) is easy to express, whereas the absence of an expected argument is
not self documenting.  Rather, it is obscure because it changes the number of
argumnets passed to a function.  It may also create further special casing if
one wishes to pass a nil argument to an extern function.

This patch regularizes the AST by representing nil explicitly through actual
arguments of type dtNil.  The consistency checks that determine formal and
actual argument types match up were updated to handle the case where the formal
argument type is dtNil (in addition to any class type).


DETAILS

    compiler/main/checks.cpp
In checkFormalActualBaseTypesMatch() and formalActualTypesMatch(), modify the
test so that passing nil to an argument of type dtNil succeeds.

    compiler/passes/parallel.cpp
Do not remove ArgSymbols of type dtNil.  These are passed through to the
backend, so formal and actual argument counts line up all the way through.

    compiler/resolution/functionResolution.cpp
Remove removeNilTypeArgs().  This optimization is no longer performed in the
resolve nor parallel passes.


[tested on / (--verify)]
[reviewed by noakes]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23497 3a8e244f-b0f2-452b-bcba-4c88e055c3ca